Refactor timeaxis plotting: move logic to utils/plot.py for reuse
